+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,2 @@
-#Wed Jan 18 23:02:57 CET 2023
-gradle.version=7.5
+#Sat Jan 14 23:42:46 CET 2023
+gradle.version=7.4
