+++ app/build.gradle
@@ -81,11 +81,5 @@ dependencies {
     implementation "com.github.bumptech.glide:glide:$version_glide"
     annotationProcessor "com.github.bumptech.glide:compiler:$version_glide"
 
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.2"
-
-    implementation 'androidx.recyclerview:recyclerview:1.1.0'
-
-
-
 
 }
