+++ app/src/main/AndroidManifest.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
-
     <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
+    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.CAMERA" />
+
 
     <application
         android:allowBackup="true"
@@ -18,16 +18,7 @@
         android:theme="@style/Theme.RSDEV"
         tools:targetApi="31">
         <activity
-            android:name=".TestMessageActivity"
-            android:exported="false" />
-        <activity
-            android:name=".Test"
-            android:exported="false" />
-        <activity
-            android:name=".MessagesReceivedActivity"
-            android:exported="false" />
-        <activity
-            android:name=".RequestsSentActivity"
+            android:name=".RequestsSendedActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
@@ -41,21 +32,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".MessagesSentActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".FriendsActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".MyMessage"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
@@ -135,7 +112,7 @@
 
         <meta-data
             android:name="preloaded_fonts"
-            android:resource="@array/preloaded_fonts" /> <!-- make sure within the application tag, otherwise app will crash with XmlResourceParser errors -->
+            android:resource="@array/preloaded_fonts" /><!-- make sure within the application tag, otherwise app will crash with XmlResourceParser errors -->
         <provider
             android:name="androidx.core.content.FileProvider"
             android:authorities="com.example.rsdev.CreatePostActivity"
