+++ app/src/main/java/com/example/rsdev/Functions.kt
@@ -49,10 +49,6 @@ class Functions {
             friend_receiver.update("friends", FieldValue.arrayUnion(id_friend_sender))
         }
 
-//        fun addMessage(id_friend_sender: String, id_friend_receiver: String) {
-//            ...............
-//        }
-
 
     }
