--- app/src/main/java/com/example/rsdev/data/ReceivedMessage.kt
@@ -1,7 +1,0 @@
-package com.example.rsdev.data
-
-
-
-data class ReceivedMessage (val messageText: String, val dateMessage: String, val receiverFirstName: String, val receiverLastName: String) {
-
-}
