--- app/src/main/java/com/example/rsdev/data/ReceivedRequest.kt
@@ -1,6 +1,0 @@
-package com.example.rsdev.data
-
-data class ReceivedRequest (val id_receiver: String, val id_sender: String, val validated: Boolean) {
-
-
-}
