--- app/src/main/java/com/example/rsdev/data/SentMessage.kt
@@ -1,5 +1,0 @@
-package com.example.rsdev.data
-
-data class SentMessage (val messageText: String, val dateMessage: String, val senderFirstName: String, val senderLastName: String) {
-
-}
