+++ app/src/main/res/layout/activity_profile.xml
@@ -111,61 +111,6 @@
         </TextView>
     </LinearLayout>
 
-    <LinearLayout
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintVertical_bias=".74"
-        app:layout_constraintHorizontal_bias=".09"
-        android:layout_width="wrap_content"
-        android:layout_height="50dp">
-        <ImageView
-            android:src="@drawable/message4"
-            android:layout_width="40dp"
-            android:layout_height="40dp"/>
-        <TextView
-            android:id="@+id/sended_messages"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="26dp"
-            android:text="Messages envoyés"
-            android:onClick="received_requests">
-        </TextView>
-    </LinearLayout>
-
-    <LinearLayout
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintVertical_bias=".82"
-        app:layout_constraintHorizontal_bias=".09"
-        android:layout_width="wrap_content"
-        android:layout_height="50dp">
-        <ImageView
-            android:src="@drawable/message4"
-            android:layout_width="40dp"
-            android:layout_height="40dp"/>
-        <TextView
-            android:id="@+id/received_messages"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="26dp"
-            android:text="Messages reçus"
-            android:onClick="received_requests">
-        </TextView>
-
-    </LinearLayout>
-
